fedimint_bitcoind
Module feerate_source

Structs

FetchJson

Constants

FEERATE_SOURCE_MAX_FEERATE_SATS_PER_VB 🔒

A feerate that we don’t expect to ever happen in practice, that we are going to reject from a source to help catching mistakes and misconfigurations.

FEERATE_SOURCE_MIN_FEERATE_SATS_PER_VB 🔒

Like FEERATE_SOURCE_MAX_FEERATE_SATS_PER_VB, but minimum one we accept

Traits

FeeRateSource