Trait fedimint_bitcoind::IBitcoindRpc

pub trait IBitcoindRpc: Debug + Send + Sync {
    fn get_network<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<Network>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_block_height<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_block_hash<'life0, 'async_trait>(
        &'life0 self,
        height: u64
    ) -> Pin<Box<dyn Future<Output = Result<BlockHash>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn get_block<'life0, 'life1, 'async_trait>(
        &'life0 self,
        hash: &'life1 BlockHash
    ) -> Pin<Box<dyn Future<Output = Result<Block>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn get_fee_rate<'life0, 'async_trait>(
        &'life0 self,
        confirmation_target: u16
    ) -> Pin<Box<dyn Future<Output = Result<Option<Feerate>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn submit_transaction<'life0, 'async_trait>(
        &'life0 self,
        transaction: Transaction
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;

    fn backend_type(&self) -> BitcoinRpcBackendType { ... }
    fn was_transaction_confirmed_in<'life0, 'life1, 'async_trait>(
        &'life0 self,
        _transaction: &'life1 Transaction,
        _height: u64
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
    { ... }
}

Trait that allows interacting with the Bitcoin blockchain

Functions may panic if if the bitcoind node is not reachable.

Required Methods

fn get_network<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<Network>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns the Bitcoin network the node is connected to

fn get_block_height<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = Result<u64>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns the current block height

fn get_block_hash<'life0, 'async_trait>(
&'life0 self,
height: u64
) -> Pin<Box<dyn Future<Output = Result<BlockHash>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Returns the block hash at a given height

Panics

If the node does not know a block for that height. Make sure to only query blocks of a height less or equal to the one returned by Self::get_block_height.

While there is a corner case that the blockchain shrinks between these two calls (through on average heavier blocks on a fork) this is prevented by only querying hashes for blocks tailing the chain tip by a certain number of blocks.

fn get_block<'life0, 'life1, 'async_trait>(
&'life0 self,
hash: &'life1 BlockHash
) -> Pin<Box<dyn Future<Output = Result<Block>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Returns the block with the given hash

Panics

If the block doesn’t exist.

fn get_fee_rate<'life0, 'async_trait>(
&'life0 self,
confirmation_target: u16
) -> Pin<Box<dyn Future<Output = Result<Option<Feerate>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Estimates the fee rate for a given confirmation target. Make sure that all federation members use the same algorithm to avoid widely diverging results. If the node is not ready yet to return a fee rate estimation this function returns None.

fn submit_transaction<'life0, 'async_trait>(
&'life0 self,
transaction: Transaction
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Submits a transaction to the Bitcoin network

Provided Methods

fn backend_type(&self) -> BitcoinRpcBackendType

true if it’s real-bitcoin (not electrum) backend and thus supports get_block call

This is a bit of a workaround to support electrum.

fn was_transaction_confirmed_in<'life0, 'life1, 'async_trait>(
&'life0 self,
_transaction: &'life1 Transaction,
_height: u64
) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Check if a transaction was included in a given (only electrum)